refactored temperature limits to be floats instead of integers

+++ api_server/forecast.py
@@ -6,8 +6,8 @@ from datetime import datetime
 class ForecastData:
     lon: float
     lat: float
-    high_temp: int
-    low_temp: int
+    high_temp: float
+    low_temp: float
     exceeds_limits: bool
     timestamp: str
     forecast_id: int = 0
